First working mouse-based version
Added the Sceitse function, which takes a canvas, and handles the drawing on
that canvas. It's taken from the example here:
http://www.williammalone.com/articles/create-html5-canvas-javascript-drawing-app/
with an undo facility and the ability to serialize/deserialize the images
added.

Next comes the touch-event handling. That's somewhat awkward, as determining
whether touch events are supported is not especially trivial when you want to
use Firefox's or Chrome's "Responsive Design" modes. But I have no doubt it's
easy once you know how.